adding functionality of squaring floating point number

+++ Square.java
@@ -1,17 +1,22 @@
+// PROBLEM STATEMENT: Program to calculate the square of a number (integer | floating point number | negative number) .
+// AUTHOR: Gaurav Garje
+
 import java.util.Scanner;
 
 public class Square {
-    static int squareCalculator(int n){
+
+    // function to calculate square.
+    static double squareCalculator(double n){
         return (n*n);
     }
     public static void main(String[] args) {
-        Scanner scan = new Scanner(System.in);
+        Scanner scan = new Scanner(System.in);      
         System.out.println("Enter a number: ");
-        int n = scan.nextInt();
+        double n = scan.nextDouble();     // Scanning a number.
 
-        System.out.println("The square of " + n + " is = " + squareCalculator(n));
+        System.out.println("The square of " + n + " is = " + squareCalculator(n));      // Calling the function to calculate square.
 
-        scan.close();
+        scan.close();   // to avoid memory leak, closing the scanner stream.
     }
 }
 
@@ -23,3 +28,6 @@ public class Square {
 // Enter a number: 
 // -5
 // The square of -5 is = 25
+// Enter a number: 
+// 7.7
+// The square of 7.7 is = 59.290000000000006
